+++ src/assets/js/calc_config.js
@@ -8,8 +8,8 @@ var CalcConfig = {
     configDefault: {
         latestVersion: {
             major: 1,
-            minor: 0,
-            revision: 2
+            minor: 1,
+            revision: 0
         },
         compact_item_modals: false
     },